add to favorites

+++ app/Http/Controllers/MediaController.php
@@ -4,14 +4,13 @@ namespace App\Http\Controllers;
 
 use App\Models\Media;
 use App\Service\ControllerSettings;
+use Auth;
 use Exception;
-use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
-use Illuminate\Foundation\Application;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
-use Psy\Util\Str;
+use Illuminate\Support\Str;
 use ReflectionClass;
 
 #[ControllerSettings(model: Media::class)]
@@ -24,14 +23,22 @@ class MediaController extends Controller
         $this->model = $this->getModel();
     }
 
-    public function adminIndex(): Factory|Application|View|\Illuminate\Contracts\Foundation\Application
+    public function adminIndex(): RedirectResponse|View
     {
+        if (Auth::user()->roles()->get()->first()->name !== 'ADMIN') {
+            return redirect()->route('app_home')->with('error', "Can't go there, admins only");
+        }
+
         $media = $this->model::all();
         return view('media.adminIndex', compact('media'));
     }
 
-    public function adminCreate(): Factory|Application|View|\Illuminate\Contracts\Foundation\Application
+    public function adminCreate(): RedirectResponse|View
     {
+        if (Auth::user()->roles()->get()->first()->name !== 'ADMIN') {
+            return redirect()->route('app_home')->with('error', "Can't go there, admins only");
+        }
+
         $media = Media::all();
 
         return view('media.adminCreate', compact('media'));
@@ -39,6 +46,10 @@ class MediaController extends Controller
 
     public function adminStore(Request $request): RedirectResponse
     {
+        if (Auth::user()->roles()->get()->first()->name !== 'ADMIN') {
+            return redirect()->route('app_home')->with('error', "Can't go there, admins only");
+        }
+
         if ($request->hasFile('media') && $request->file('media')->isValid()) {
             $newMedia = $request->file('media');
             $newPath = $newMedia->store('public/images');
@@ -56,22 +67,34 @@ class MediaController extends Controller
         return redirect()->route('admin_media_create')->with('error', 'Media not created.');
     }
 
-    public function adminShow(Media $media): Factory|Application|View|\Illuminate\Contracts\Foundation\Application
+    public function adminShow(Media $media): RedirectResponse|View
     {
+        if (Auth::user()->roles()->get()->first()->name !== 'ADMIN') {
+            return redirect()->route('app_home')->with('error', "Can't go there, admins only");
+        }
+
         return view('media.adminShow', compact('media'));
     }
 
-    public function adminEdit(Media $media): Factory|Application|View|\Illuminate\Contracts\Foundation\Application
+    public function adminEdit(Media $media): RedirectResponse|View
     {
+        if (Auth::user()->roles()->get()->first()->name !== 'ADMIN') {
+            return redirect()->route('app_home')->with('error', "Can't go there, admins only");
+        }
+
         return view('media.adminEdit', compact('media'));
     }
 
     public function adminUpdate(Request $request, Media $media): RedirectResponse
     {
+        if (Auth::user()->roles()->get()->first()->name !== 'ADMIN') {
+            return redirect()->route('app_home')->with('error', "Can't go there, admins only");
+        }
+
         try {
             if ($request->hasFile('media') && $request->file('media')->isValid()) {
                 $newMedia = $request->file('media');
-                $newPath = $newMedia->store('images');
+                $newPath = $newMedia->store('public/images');
                 $newSize = $newMedia->getSize();
                 $newAlt = fake()->words(3, true);
 
@@ -89,10 +112,14 @@ class MediaController extends Controller
 
     public function adminDestroy(Media $media, Request $request): RedirectResponse
     {
+        if (Auth::user()->roles()->get()->first()->name !== 'ADMIN') {
+            return redirect()->route('app_home')->with('error', "Can't go there, admins only");
+        }
+
         $media->delete();
 
-        if (\Illuminate\Support\Str::contains($request->headers->get('referer'), ['lodging', 'show'], true)) {
-            return redirect()->back();
+        if (Str::contains($request->headers->get('referer'), ['lodging', 'show'], true)) {
+            return redirect()->back()->with('success', 'Media deleted from lodging successfully');
         }
 
         return redirect()->route('admin_media_index')->with('success', 'Media deleted successfully.');
